+++ README.md
@@ -8,5 +8,5 @@ vue的渐进式练习
 4，npm install vuex -S：安装vue的状态管理插件vuex
 
 #第二天：集成并使用vue-router（基础知识点参照vue-router-base.txt）  
-1，router/index.js下，export基础对象{routes: []}
-2，在入口文件main.js中实例化路由，Vue.use(Router)必须在实例化之前
+1，router/index.js下，export基础对象{routes: []}  
+2，在入口文件main.js中实例化路由，Vue.use(Router)必须在实例化之前
