+++ src/router/index.js
@@ -3,37 +3,40 @@ import menus from '@/page/menus'
 import TakeoutIndex from '@/page/takeout'
 import TakeoutProduct from '@/page/takeout/product'
 
-export default {
-  routes: [
+//四个业务的路由
+const businessRoutes = {
+  path: '/',
+  component: layout,
+  children: [
     {
-      path: '/',
-      component: layout,
-      children: [
+      path: 'takeout',
+      component: menus,
+      beforeEnter(to, from, next) {
+        console.log("change lastlogintype");
+        next();
+      },
+      children:[
         {
-          path: 'takeout',
-          component: menus,
-          children:[
-            {
-              path: '',
-              name: 'takeoutInfo',
-              component: {
-                template: `<div><h1>外卖业务模块，请选择菜单查看具体功能</h1></div>`
-              }
-            },
-            {
-              path: 'index',
-              name: 'index',
-              component: TakeoutIndex
-            },{
-              path: 'product',
-              name: 'product',
-              component: TakeoutProduct
-            }
-          ]
+          path: '',
+          rediret: 'index'
+        },
+        {
+          path: 'index',
+          name: 'index',
+          component: TakeoutIndex,
+        },{
+          path: 'product',
+          name: 'product',
+          component: TakeoutProduct
         }
       ]
     }
-    
+  ]
+}
 
+export default {
+  routes: [
+    businessRoutes
+    
   ]
 }
