+++ static/css/global.css
@@ -1,4 +1,10 @@
-body, p {
+body{
+  margin: 0;
+  color: #333333;
+  background: #f1f1f1;
+}
+
+p {
   margin: 0;
 }
